+++ cicerone/spec/__init__.py
@@ -1,0 +1,19 @@
+"""OpenAPI specification models and utilities."""
+
+from cicerone.spec.components import Components
+from cicerone.spec.openapi_spec import OpenAPISpec
+from cicerone.spec.operation import Operation
+from cicerone.spec.path_item import PathItem
+from cicerone.spec.paths import Paths
+from cicerone.spec.schema import Schema
+from cicerone.spec.version import Version
+
+__all__ = [
+    "Components",
+    "OpenAPISpec",
+    "Operation",
+    "PathItem",
+    "Paths",
+    "Schema",
+    "Version",
+]